refactor: interfaces - disect  interfaces, routes -  creater seperate folder for routers and put all router in them
