+++ backend/backend/backend-src/db/events.ts
@@ -5,8 +5,8 @@
 import {DynamoDBClient, QueryCommand} from '@aws-sdk/client-dynamodb'
 import {marshall, unmarshall} from '@aws-sdk/util-dynamodb'
 
-import {DB} from '../../constants'
-import {tickerToPackageName} from '../utils'
+import {DB} from '../constants'
+import {tickerToPackageName} from '../lib/utils'
 
 const DB_CLIENT = new DynamoDBClient()
