+++ backend/backend/backend-src/interfaces/relation.ts
@@ -1,0 +1,20 @@
+/**
+ * @file related Interfaces and types.
+ */
+
+export interface IRelatedItem {
+    label: string
+    address: string
+  }
+  export interface IRelationCreate {
+    label: string
+    address: string
+  }
+  export interface IRelationDelete {
+    label: string
+  }
+  
+  export interface IRelationRename {
+    label: string
+  }
+
