+++ backend/backend/backend-src/interfaces/valid.ts
@@ -1,0 +1,4 @@
+export interface IValid {
+  error: string
+  data?: object
+}
