+++ backend/backend/backend-src/lib/index.ts
@@ -1,0 +1,38 @@
+import {
+  handleCreate,
+  handleCancel,
+  handleIconUrlRequest,
+  handlePublished,
+  handleGetFilteredPackages,
+  handleGetPackages,
+} from './packages'
+
+import {handleGetRelations, handleCreateRelation, handleRenameRelation, handleDeleteRelation} from './relations'
+
+import {handleGetPackageEvents, handleGetAddressEvents, handleGetBalances, handleGetAllocations} from './events'
+import {handleGetTxVol} from './txvol'
+import {handleGetHoldings} from './holdings'
+
+export {
+  //packages
+  handleCreate,
+  handleCancel,
+  handleIconUrlRequest,
+  handlePublished,
+  handleGetFilteredPackages,
+  handleGetPackages,
+  // relations
+  handleGetRelations,
+  handleCreateRelation,
+  handleRenameRelation,
+  handleDeleteRelation,
+  // events
+  handleGetPackageEvents,
+  handleGetAddressEvents,
+  handleGetBalances,
+  handleGetAllocations,
+  //tx volume
+  handleGetTxVol,
+  // holdings
+  handleGetHoldings,
+}
