+++ backend/backend/backend-src/lib/packages.ts
@@ -8,70 +8,12 @@ import {Zip} from 'zip-lib'
 
 import * as child from 'child_process'
 import fs from 'fs'
-import {TOKEN_SUPPLY_PATH} from './../constants'
 import * as Checks from './checks'
-import * as dbPackages from './db/packages'
-import {ErrorType, invalidAddressErrorDetail, S3MoneyError} from './error'
-import * as IFace from './interfaces'
+import * as dbPackages from '../db/packages'
+import {ErrorType, invalidAddressErrorDetail, S3MoneyError} from '../interfaces/error'
+import * as IFace from '../interfaces'
 import * as storage from './storage'
 import {tickerToPackageName} from './utils'
-
-const CWD = process.cwd()
-const TOKEN_PATH = `${CWD}/${TOKEN_SUPPLY_PATH}`
-const WORK_DIR = process.env.WORK_DIR || `${CWD}/contracts`
-
-export function createPackagesRouter(): Router {
-  const router = Router()
-  router.use((req: Request, res, next) => {
-    req.tokenPath = TOKEN_PATH
-    req.workDir = WORK_DIR
-    next()
-  })
-  router.post('/create', async (req, res, next) => {
-    try {
-      await handleCreate(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.post('/cancel', async (req, res, next) => {
-    try {
-      await handleCancel(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.post('/published', async (req, res, next) => {
-    try {
-      await handlePublished(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.post('/generateIconURL', async (req, res, next) => {
-    try {
-      await handleIconUrlRequest(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.get('/:address', async (req, res, next) => {
-    try {
-      await handleGetPackages(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.get('/:address/:param', async (req, res, next) => {
-    try {
-      await handleGetFilteredPackages(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  return router
-}
 
 /**
  * Create a package.
