+++ backend/backend/backend-src/lib/relations.ts
@@ -2,40 +2,10 @@
  * @file Handlers related to wallet relationships.
  */
 
-import {Request, Response, Router} from 'express'
+import {Request, Response} from 'express'
 import * as Checks from './checks'
-import * as dbRelations from './db/relations'
-import * as IFace from './interfaces'
-
-export function createRelationsRouter(): Router {
-  const router = Router()
-
-  router.get('/:pkgAddress', async (req, res, next) => {
-    try {
-      await handleGetRelations(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  router.post('/:pkgAddress', async (req, res, next) => {
-    try {
-      await handleCreateRelation(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-  // implemented, but not part of the current deliverable
-  // router.delete('/related/:pkgAddress/:slug', handleDeleteRelation)
-  router.patch('/:pkgAddress/:slug', async (req, res, next) => {
-    try {
-      await handleRenameRelation(req, res)
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  return router
-}
+import * as dbRelations from '../db/relations'
+import * as IFace from '../interfaces'
 
 /**
  * Handler that returns a list of relations for a package.
