+++ backend/backend/backend-src/routes/index.ts
@@ -1,0 +1,7 @@
+import {createPackagesRouter} from './packages-api'
+import {createRelationsRouter} from './relations-api'
+import {createEventsRouter} from './events-api'
+import {createTxVolRouter} from './txvol-api'
+import {createHoldingsRouter} from './holdings-api'
+
+export {createPackagesRouter, createRelationsRouter, createEventsRouter, createTxVolRouter, createHoldingsRouter}
